feat: 싱글턴 적용-3 #1
